# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenReturn API
  version: 0.1.0
  description: Reference REST API for the OpenReturn protocol.
servers:
  - url: http://localhost:4000
security:
  - bearerAuth: []
...
paths:
...
  /oauth/token:
    post:
      tags: [OAuth]
      summary: Issue OAuth client credentials token
      security: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                client_id:
                  type: string
                scope:
                  type: string
      responses:
        "200":
          description: Token response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthTokenResponse"
  /oauth/delegate:
    post:
      tags: [OAuth]
      summary: Exchange a subject token for a delegated agent token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenDelegationRequest"
      responses:
        "200":
          description: Delegated token response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OAuthTokenResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /orders/{id}:
    get:
      tags: [Orders]
      summary: Lookup an order through the configured commerce adapter
      parameters:
        - $ref: "#/components/parameters/IdPath"
        - name: email
          in: query
          schema:
            type: string
            format: email
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                type: object
                properties:
                  order:
                    $ref: "#/components/schemas/Order"
        "404":
          $ref: "#/components/responses/Error"
        "503":
          $ref: "#/components/responses/Error"
  /returns:
    get:
      tags: [Returns]
      summary: List returns
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ReturnState"
        - name: email
          in: query
          schema:
            type: string
            format: email
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 500
      responses:
        "200":
          description: Return list
          content:
            application/json:
              schema:
                type: object
                properties:
                  returns:
                    type: array
                    items:
                      $ref: "#/components/schemas/OpenReturnRecord"
        "400":
          $ref: "#/components/responses/Error"
    post:
      tags: [Returns]
      summary: Initiate a return
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateReturnRequest"
      responses:
        "201":
          description: Return created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnResponse"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /returns/{id}:
    get:
      tags: [Returns]
      summary: Get a return
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Return
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnResponse"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [Returns]
      summary: Update return lifecycle data
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateReturnRequest"
      responses:
        "200":
          description: Updated return
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /returns/{id}/exchange:
    post:
      tags: [Returns]
      summary: Select exchange products
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectExchangeRequest"
      responses:
        "200":
          description: Updated return
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /returns/{id}/carrier:
    post:
      tags: [Returns]
      summary: Select carrier and generate label
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectCarrierRequest"
      responses:
        "200":
          description: Updated return with label
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /returns/{id}/label:
    get:
      tags: [Labels]
      summary: Get shipping label metadata
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Shipping label
          content:
            application/json:
              schema:
                type: object
                required: [label]
                properties:
                  label:
                    $ref: "#/components/schemas/ShippingLabel"
        "404":
          $ref: "#/components/responses/Error"
  /labels/{carrier}/{serviceLevel}/{trackingNumber}:
    get:
      tags: [Labels]
      summary: Download a mock label PDF
      security: []
      parameters:
        - name: carrier
          in: path
          required: true
          schema:
            type: string
        - name: serviceLevel
          in: path
          required: true
          schema:
            type: string
        - name: trackingNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF label
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /labels/{carrier}/{trackingNumber}:
    get:
      tags: [Labels]
      summary: Download a mock label PDF using the legacy label path
      security: []
      parameters:
        - name: carrier
          in: path
          required: true
          schema:
            type: string
        - name: trackingNumber
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF label
          content:
            application/pdf:
              schema:
                type: string
                format: binary
  /returns/{id}/track:
    post:
      tags: [Returns]
      summary: Add tracking update
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddTrackingRequest"
      responses:
        "200":
          description: Updated return
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReturnResponse"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /returns/{id}/events:
    get:
      tags: [Returns]
      summary: Get return event history
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Event history
          content:
            application/json:
              schema:
                type: object
                required: [events]
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReturnEvent"
        "404":
          $ref: "#/components/responses/Error"
  /webhooks:
    post:
      tags: [Webhooks]
      summary: Receive carrier or commerce webhook events
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookEvent"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                required: [accepted]
                properties:
                  accepted:
                    type: boolean
                  return:
                    oneOf:
                      - $ref: "#/components/schemas/OpenReturnRecord"
                      - type: "null"
        "400":
          $ref: "#/components/responses/Error"
components:
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ReturnState:
      type: string
      enum: [initiated, label_generated, shipped, in_transit, delivered, inspection, approved, rejected, refunded, exchanged, completed]
    ResolutionType:
      type: string
      enum: [refund, exchange, store_credit, coupon_code]
    ReasonCode:
      type: string
      enum: [defect, size, not_as_described, wrong_item, arrived_late, damaged_in_transit, duplicate_order, changed_mind, unwanted, other]
    CarrierCode:
      type: string
      enum: [postnl, dhl, ups, dpd, budbee]
    TrackingStatus:
      type: string
      enum: [label_created, accepted, in_transit, out_for_delivery, delivered, exception]
    Money:
      type: object
      required: [amount, currency]
      properties:
        amount:
          type: integer
        currency:
          type: string
          minLength: 3
          maxLength: 3
    Address:
      type: object
      required: [line1, city, postalCode, countryCode]
      properties:
        name: { type: string }
        line1: { type: string }
        line2: { type: string }
        city: { type: string }
        region: { type: string }
        postalCode: { type: string }
        countryCode: { type: string }
    Customer:
      type: object
      required: [email]
      properties:
        id: { type: string }
        email: { type: string, format: email }
        name: { type: string }
        phone: { type: string }
        shippingAddress:
          $ref: "#/components/schemas/Address"
    OrderItem:
      type: object
      required: [id, sku, name, quantity, unitPrice]
      properties:
        id: { type: string }
        sku: { type: string }
        name: { type: string }
        quantity: { type: integer, minimum: 1 }
        unitPrice:
          $ref: "#/components/schemas/Money"
        imageUrl: { type: string }
        attributes:
          type: object
          additionalProperties:
            type: string
    Order:
      type: object
      required: [id, customer, items, total, placedAt]
      properties:
        id: { type: string }
        externalOrderId: { type: string }
        customer:
          $ref: "#/components/schemas/Customer"
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        total:
          $ref: "#/components/schemas/Money"
        placedAt: { type: string, format: date-time }
        platform: { type: string }
    ReturnReason:
      type: object
      required: [code]
      properties:
        code:
          $ref: "#/components/schemas/ReasonCode"
        note: { type: string }
        evidenceUrls:
          type: array
          items: { type: string, format: uri }
    ReturnItem:
      type: object
      required: [orderItemId, sku, name, quantity, reason]
      properties:
        orderItemId: { type: string }
        sku: { type: string }
        name: { type: string }
        quantity: { type: integer, minimum: 1 }
        reason:
          $ref: "#/components/schemas/ReturnReason"
        condition:
          type: string
          enum: [unopened, new, used, defective, damaged]
    ExchangeItem:
      type: object
      required: [originalOrderItemId, replacementSku, replacementName, quantity]
      properties:
        originalOrderItemId: { type: string }
        replacementSku: { type: string }
        replacementName: { type: string }
        quantity: { type: integer, minimum: 1 }
        attributes: { type: object }
        priceDifference:
          $ref: "#/components/schemas/Money"
    ExchangeSelection:
      type: object
      required: [requestedItems, selectedAt, status]
      properties:
        requestedItems:
          type: array
          items:
            $ref: "#/components/schemas/ExchangeItem"
        selectedAt: { type: string, format: date-time }
        status:
          type: string
          enum: [pending, reserved, fulfilled, cancelled]
    InitiateReturnRequest:
      type: object
      required: [orderId, customer, items, requestedResolution]
      properties:
        orderId: { type: string }
        externalOrderId: { type: string }
        customer:
          $ref: "#/components/schemas/Customer"
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ReturnItem"
        requestedResolution:
          $ref: "#/components/schemas/ResolutionType"
        returnMethod: { type: string }
        metadata: { type: object }
    UpdateReturnRequest:
      type: object
      minProperties: 1
      properties:
        status:
          $ref: "#/components/schemas/ReturnState"
        inspection:
          $ref: "#/components/schemas/InspectionResult"
        refund:
          $ref: "#/components/schemas/RefundResult"
        storeCredit:
          $ref: "#/components/schemas/StoreCreditResult"
        couponCode:
          $ref: "#/components/schemas/CouponCodeResult"
        metadata:
          type: object
          additionalProperties: true
    SelectExchangeRequest:
      type: object
      required: [requestedItems]
      properties:
        requestedItems:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ExchangeItem"
    SelectCarrierRequest:
      type: object
      required: [carrier]
      properties:
        carrier:
          $ref: "#/components/schemas/CarrierCode"
        serviceLevel: { type: string }
        dropoffPointId: { type: string }
        pickupWindow:
          type: object
          required: [startsAt, endsAt]
          properties:
            startsAt: { type: string, format: date-time }
            endsAt: { type: string, format: date-time }
        shipFrom:
          $ref: "#/components/schemas/Address"
        shipTo:
          $ref: "#/components/schemas/Address"
    ShippingLabel:
      type: object
      required: [id, carrier, trackingNumber, labelUrl, format, expiresAt, createdAt]
      properties:
        id: { type: string }
        carrier: { type: string }
        trackingNumber: { type: string }
        labelUrl: { type: string, format: uri }
        format:
          type: string
          enum: [pdf, png, zpl]
        expiresAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    TrackingEvent:
      type: object
      required: [id, carrier, trackingNumber, status, occurredAt]
      properties:
        id: { type: string }
        carrier: { type: string }
        trackingNumber: { type: string }
        status:
          $ref: "#/components/schemas/TrackingStatus"
        occurredAt: { type: string, format: date-time }
        location: { type: string }
        description: { type: string }
    InspectionResult:
      type: object
      required: [inspectedAt, accepted]
      properties:
        inspectedAt: { type: string, format: date-time }
        inspectorId: { type: string }
        accepted: { type: boolean }
        notes: { type: string }
        itemConditions:
          type: object
          additionalProperties:
            type: string
            enum: [unopened, new, used, defective, damaged]
    RefundResult:
      type: object
      required: [amount, provider, transactionId, processedAt]
      properties:
        amount:
          $ref: "#/components/schemas/Money"
        provider: { type: string }
        transactionId: { type: string }
        processedAt: { type: string, format: date-time }
    StoreCreditResult:
      type: object
      required: [amount, code, issuedAt]
      properties:
        amount:
          $ref: "#/components/schemas/Money"
        code: { type: string }
        expiresAt: { type: string, format: date-time }
        issuedAt: { type: string, format: date-time }
    CouponCodeResult:
      type: object
      required: [code, issuedAt]
      properties:
        code: { type: string }
        amount:
          $ref: "#/components/schemas/Money"
        percentage: { type: number, exclusiveMinimum: 0, maximum: 100 }
        expiresAt: { type: string, format: date-time }
        issuedAt: { type: string, format: date-time }
    ReturnEvent:
      type: object
      required: [id, returnId, type, state, message, createdAt]
      properties:
        id: { type: string }
        returnId: { type: string }
        type: { type: string }
        state:
          $ref: "#/components/schemas/ReturnState"
        message: { type: string }
        actor: { type: string }
        data:
          type: object
          additionalProperties: true
        createdAt: { type: string, format: date-time }
    OpenReturnRecord:
      type: object
      required: [id, orderId, customer, status, requestedResolution, reasonCodes, items, returnMethod, tracking, events, createdAt, updatedAt]
      properties:
        id: { type: string }
        orderId: { type: string }
        externalOrderId: { type: string }
        customer:
          $ref: "#/components/schemas/Customer"
        status:
          $ref: "#/components/schemas/ReturnState"
        requestedResolution:
          $ref: "#/components/schemas/ResolutionType"
        reasonCodes:
          type: array
          items:
            $ref: "#/components/schemas/ReasonCode"
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReturnItem"
        returnMethod: { type: string }
        exchange:
          $ref: "#/components/schemas/ExchangeSelection"
        carrier:
          $ref: "#/components/schemas/SelectCarrierRequest"
        label:
          $ref: "#/components/schemas/ShippingLabel"
        tracking:
          type: array
          items:
            $ref: "#/components/schemas/TrackingEvent"
        events:
          type: array
          items:
            $ref: "#/components/schemas/ReturnEvent"
        inspection:
          $ref: "#/components/schemas/InspectionResult"
        refund:
          $ref: "#/components/schemas/RefundResult"
        storeCredit:
          $ref: "#/components/schemas/StoreCreditResult"
        couponCode:
          $ref: "#/components/schemas/CouponCodeResult"
        metadata:
          type: object
          additionalProperties: true
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        completedAt: { type: string, format: date-time }
    AddTrackingRequest:
      type: object
      required: [status]
      properties:
        trackingNumber: { type: string }
        status:
          $ref: "#/components/schemas/TrackingStatus"
        occurredAt: { type: string, format: date-time }
        location: { type: string }
        description: { type: string }
    WebhookEvent:
      type: object
      required: [source, type, data]
      properties:
        id: { type: string }
        source: { type: string }
        type: { type: string }
        returnId: { type: string }
        trackingNumber: { type: string }
        data:
          type: object
          additionalProperties: true
        occurredAt: { type: string, format: date-time }
    TokenDelegationRequest:
      type: object
      required: [subjectToken, actor, scope]
      properties:
        subjectToken: { type: string }
        actor: { type: string }
        scope: { type: string }
        audience: { type: string }
    OAuthTokenResponse:
      type: object
      required: [access_token, token_type, expires_in, scope]
      properties:
        access_token: { type: string }
        token_type: { type: string, enum: [Bearer] }
        expires_in: { type: integer }
        scope: { type: string }
        delegated_subject: { type: string }
...
    ReturnResponse:
      type: object
      required: [return]
      properties:
        return:
          $ref: "#/components/schemas/OpenReturnRecord"
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details: {}